separated move_player.c into main.c and move_player.c
